Screenshots
In the zip folder there are stored, in an organized way, all the necessary screenshots per instance, for every outcome of the test completion, as a proof of success. You will find screenshots of the functionalities required by the test, either for the custom component, or for the custom module.